cleanup: dissolve the fake CMapStringToOb map views + type EnterModalUI's message
Follow-up to RELOC-Multi, per user direction (finish the deferred items, don't defer).

1) EnterModalUI's arg is the modal TEXT, not an id. Disasm-proven: CGruntzMgr::
   EnterModalUI@0x8ef10 forwards it to CGameApp::RunModal@0x80c00, whose body does
   strlen + strcpy of arg1 into the g_644ea0 message buffer and then DialogBoxParamA's
   the message dialog. Retyped the whole chain (WAP32::CGameApp::RunModal, CGruntzMgr::
   EnterModalUI, CGameRegistry::EnterModalUI) from `i32` to `const char*`, which deletes
   six (i32)/(i32)(const char*) casts at the call sites. Folded the two fake aliases of
   the SAME rva onto it: CNetGameMgr::ShowModal and MultiStartDlgRoster's CGameSettings::
   ShowModal (both @0x8ef10) are gone; their 4 reporters call the canonical method.

2) The `((CMapStringToOb*)&x)->Lookup(k, (CObject*&)v)` spelling is a DOUBLE lie and is
   now gone tree-wide: an object-cast because the map member was typed as an invented
   empty view, plus a reinterpret of a `T*&` as a `CObject*&`. Dissolved SIXTEEN phantom
   map classes - all empty shells whose single "method" was a fake alias of the MFC
   library ?Lookup@CMapStringToOb@@QBEHPBDAAPAVCObject@@@Z at 0x1b8438, so every call
   through them bound to nothing: CSndFinder, CSpriteHashTable, CMiCueMap, CSbiLookupMap,
   RockSndTable, CEntranceHashTable, CombatCueMap, CProjSpriteMap, CBlitLabelMap,
   CLogicMap, CRegBute, CResMapStr, GruntSoundMap, HazStrMap, HbSndTable, Map_bdd0,
   SbiSndTable, CChatMap, CEffMap. Their members are now the real CMapStringToOb (its
   0x1c bytes exactly fill CSndHost +0x10..+0x2c), callers use the container's own
   CObject* out-param, and the calls link against the library symbol.
   Removed: 74 object-casts, 48 (CObject*&) reinterprets, 2 (LeafCue**)& arg-casts.

   The documented "~60 pure-Win32 TU C1189 wall" that blocked this was MEASURED and is
   false: only TWO Win32-umbrella TUs reach SoundCue.h transitively (LightFxRender.cpp,
   MainMenuBuilder.cpp). Both switched to <Mfc.h> (the superset) with zero C1189.

   NOT touched: the 0x1b8008 map family - that is a DIFFERENT library class
   (?Lookup@CMapStringToPtr@@QBEHPBDAAPAX@Z), so CSbiConfigMap et al. stay as they are.

Build: all FATAL gates pass. Exact count UP 2263 -> 2265; fuzzy 71.94 -> 71.95; MAX %
high-water unchanged. 27 per-function regressions accepted, mechanism understood: the
honest form hoists a `CObject*` out-var and then downcasts into the typed local, so those
frames carry ONE extra stack slot vs retail (which keeps a single slot). Follow-up to
recover them: collapse to a single `CObject*` local with the downcast at each USE site
(same one slot as retail) instead of a second variable. Reverting to the fake spelling is
not an option - it is the thing being removed.

(cherry picked from commit 79c2e6870118712ea3e7af0666ac859adfeb0393)

+++ src/Gruntz/SerializeSyncMarker.cpp
@@ -39,7 +39,7 @@ i32 SerializeSyncMarker(CSerialArchive* arc, i32 mode, const char* name, i32 lin
         arc->Read(&readVal, 4);
         if (readVal != g_serialCounter + 0x1234666) {
             wsprintfA(g_629a50, "save/load out of sync at %s, %d", name, line);
-            g_gameReg->EnterModalUI((i32)g_629a50);
+            g_gameReg->EnterModalUI(g_629a50);
             return 0;
         }
     }
